pam: Move pam_test into internal packages
