+++ .github/workflows/qa.yaml
@@ -156,7 +156,7 @@ jobs:
           CGO_CFLAGS: "-O0 -g3 -fno-omit-frame-pointer"
         run: |
           # Use `-dwarflocationlists` to give ASAN a better time to unwind the stack trace
-          go test -asan -gcflags="-dwarflocationlists=true" ./pam/pam_test ./pam/internal/gdm/
+          go test -C ./pam/internal -asan -gcflags="-dwarflocationlists=true" ./...
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
         with:
