+++ pam/internal/gdm/extension_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/msteinert/pam/v2"
 	"github.com/stretchr/testify/require"
-	"github.com/ubuntu/authd/pam/pam_test"
+	"github.com/ubuntu/authd/pam/internal/pam_test"
 )
 
 func TestMain(t *testing.T) {
